+++ documentation/src/components/github-floating-cta/index.tsx
@@ -23,11 +23,11 @@ const GithubFloatingCta: FC = () => {
     if (isClosed) return null;
 
     return (
-        <div className="font-montserrat shadow-githubFloatingCta sticky bottom-0 mx-auto flex h-[48px] w-full items-center bg-[#2A2A42] px-3 text-xs font-bold text-white sm:bottom-[32px] sm:w-[350px] sm:rounded-3xl">
+        <div className="font-montserrat shadow-githubFloatingCta fixed bottom-0 mx-auto flex h-[48px] w-full items-center bg-[#2A2A42] px-3 text-xs font-bold text-white left-0 right-0 sm:bottom-[32px] sm:w-[350px] sm:rounded-3xl z-[999]">
             <div className="ml-4">Star us on GitHub</div>
             {/* eslint-disable-next-line react/jsx-no-target-blank */}
             <a
-                className="text--no-decoration mx-auto flex items-center gap-2 text-inherit"
+                className="flex items-center gap-2 mx-auto text--no-decoration text-inherit"
                 href="https://github.com/refinedev/refine/?ref=float-cta"
                 target="_blank"
             >