+++ packages/chakra-ui/CHANGELOG.md
@@ -1,5 +1,51 @@
 # @pankod/refine-chakra-ui
 
+## 1.2.6
+
+### Patch Changes
+
+-   [#2959](https://github.com/refinedev/refine/pull/2959) [`3cd13fa5c2`](https://github.com/refinedev/refine/commit/3cd13fa5c24a38c3a14458e26a8a43567144b4ea) Thanks [@salihozdemir](https://github.com/salihozdemir)! - - Fixed error and ready pages issues that dark mode compatibility and mobile view issues.
+    -   Fixed crud components mobile view issues and spacing issues.
+    -   Update `Save` button icon button color and variant.
+    -   Rename `TextFieldComponent` export name to `TextField`.
+    -   Update `DeleteButton` color scheme.
+    -   Add missing props to `RefreshButton`'s `IconButton`.
+
+## 1.2.5
+
+### Patch Changes
+
+-   [#2959](https://github.com/refinedev/refine/pull/2959) [`3cd13fa5c2`](https://github.com/refinedev/refine/commit/3cd13fa5c24a38c3a14458e26a8a43567144b4ea) Thanks [@salihozdemir](https://github.com/salihozdemir)! - - Fixed error and ready pages issues that dark mode compatibility and mobile view issues.
+    -   Fixed crud components mobile view issues and spacing issues.
+    -   Update `Save` button icon button color and variant.
+    -   Rename `TextFieldComponent` export name to `TextField`.
+    -   Update `DeleteButton` color scheme.
+    -   Add missing props to `RefreshButton`'s `IconButton`.
+
+## 1.2.4
+
+### Patch Changes
+
+-   [#2948](https://github.com/refinedev/refine/pull/2948) [`add3da4c76`](https://github.com/refinedev/refine/commit/add3da4c76472d6c025aa3a2f0f21ecd3ab1589e) Thanks [@yildirayunlu](https://github.com/yildirayunlu)! - Fix AuthPage styles props and dark theme colors.
+
+## 1.2.3
+
+### Patch Changes
+
+-   [#2948](https://github.com/refinedev/refine/pull/2948) [`add3da4c76`](https://github.com/refinedev/refine/commit/add3da4c76472d6c025aa3a2f0f21ecd3ab1589e) Thanks [@yildirayunlu](https://github.com/yildirayunlu)! - Fix AuthPage styles props and dark theme colors.
+
+## 1.2.2
+
+### Patch Changes
+
+-   [`59d6bd5b83`](https://github.com/refinedev/refine/commit/59d6bd5b832e96304623ed8417d77a6556a41b82) Thanks [@salihozdemir](https://github.com/salihozdemir)! - Added `href` property for `BreadcrumbItem` component to be able to fix the Next.js undefined `href` error.
+
+## 1.2.1
+
+### Patch Changes
+
+-   [`59d6bd5b83`](https://github.com/refinedev/refine/commit/59d6bd5b832e96304623ed8417d77a6556a41b82) Thanks [@salihozdemir](https://github.com/salihozdemir)! - Added `href` property for `BreadcrumbItem` component to be able to fix the Next.js undefined `href` error.
+
 ## 1.2.0
 
 ### Minor Changes